Prom sample in a separate view
Having Prometheus sample baing part of the main view used for querying the data adds a lot of overhead (mostly due to the fact that for each row it invokes C function to create a sample). In most use cases you are not interested in fetching raw samples. We create an additional view that will handle Prometheus samples only.